TESTING: Add lru_sync syscall
This is NOT intended for upstream. Some of my THP tests need new pages to
have PageLRU (that is, they are immediately flushed out of the pagevec
cache), and using a syscall seems to do the job. Please let me know if you
have a better way, I don't like this.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>