tools: smmute: Add VFIO-based userspace driver
We add two drivers to smmute, alongside the kernel one:

VFIO: setup direct access to the device using VFIO. The userspace driver
mmaps the device frames, maps its buffer through VFIO_MAP/UNMAP, and
handles transactions by reading and writing into the frames. No MSI
support for the moment, as this is still work in progress in the kernel.

By assigning a device directly to the process, we introduce a big
restriction. Another process will have to wait for the driver to release
the device's group before being able to access the device.

IPC: to circumvent this restriction, introduce a server (smmuted) that
owns the device via VFIO, and multiple clients requesting mappings and
transactions. Contemporary applications based on this model would use
something like D-BUS for transport, but we keep it simple for the moment,
with UNIX sockets and a light handcrafted stream protocol. We use shm
objects to share memory buffers.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>